+++ code/Motor_Codes/Test_Motor_controls_v2/motor_controls.h
@@ -4,12 +4,12 @@
 */ 
 #ifndef MOTOR_CONTROLS_H
 #define MOTOR_CONTROLS_H
-#include <arduino.h>
+#include <Arduino.h>
 
 
 void motorSetup(byte EN1, byte IN1 ,byte IN2, byte EN2, byte EN3, byte EN4);
 
-void setSpeed(int snelheid);
+void setSpeed(uint8_t snelheid);
 void forward();
 void stop();
 void backward();
